+++ i18nilize/tests/test_cli.py
@@ -1,7 +1,9 @@
 import unittest, os, json, timeit, shutil
 from unittest.mock import patch
-from src.internationalize.helpers import delete_translation, get_json, make_translation_map, get_translation, add_language, add_update_translated_word, pull_translations
+from src.internationalize.helpers import delete_translation, get_json, make_translation_map, get_translation, add_language, add_update_translated_word
 from src.internationalize import globals
+from src.internationalize.sync_processor import pull_translations, push_translations
+from src.internationalize.diffing_processor import DiffingProcessor
 
 # Create your tests here.            
 # To test:
@@ -162,12 +164,19 @@ class TestCLI(unittest.TestCase):
         prev_token = globals.token.value
         test_token = "c84234c3-b507-4ed0-a6eb-8b10116cdef1"
         globals.token.value = test_token
+        temp_dir_path = os.path.join(self.languages_dir, "temp")
+        diff_processor = DiffingProcessor(temp_dir_path)
+        if os.path.exists(diff_processor.diff_state_root_dir):
+            shutil.rmtree(diff_processor.diff_state_root_dir)
+        diff_processor.setup()
 
         # Create temporary directories to pull translations
-        temp_dir_path = os.path.join(self.languages_dir, "temp")
         files_to_copy = ["spanish.json", "french.json"]
-        if not os.path.exists(temp_dir_path):
-            os.mkdir(temp_dir_path)
+        if os.path.exists(temp_dir_path):
+            shutil.rmtree(temp_dir_path)
+        os.mkdir(temp_dir_path)
+        
+        # Copy files into temp dir to test overwriting
         for file_name in files_to_copy:
             curr_file_path = os.path.join(self.languages_dir, file_name)
             new_file_path = os.path.join(temp_dir_path, file_name)
@@ -197,8 +206,52 @@ class TestCLI(unittest.TestCase):
             self.assertEqual(file_content, expected_file_content[file_name])
 
         # Cleanup
+        shutil.rmtree(diff_processor.diff_state_root_dir)
         shutil.rmtree(temp_dir_path)
         globals.token.value = prev_token
+
+    # def test_push_translations(self):
+    #     prev_token = globals.token.value
+    #     test_token = "c53a8272-7e50-4757-90c4-2befd33d70cf"
+    #     globals.token.value = test_token
+    #
+    #     # Create temporary directories to pull translations
+    #     temp_dir_path = os.path.join(self.languages_dir, "temp")
+    #     files_to_copy = ["spanish.json", "french.json"]
+    #     if os.path.exists(temp_dir_path):
+    #         shutil.rmtree(temp_dir_path)
+    #     os.mkdir(temp_dir_path)
+    #     for file_name in files_to_copy:
+    #         curr_file_path = os.path.join(self.languages_dir, file_name)
+    #         new_file_path = os.path.join(temp_dir_path, file_name)
+    #         shutil.copy(curr_file_path, new_file_path)
+    #
+    #     # Expected content after pulling from API
+    #     expected_file_content = {
+    #         "fr.json": {
+    #             "hello": "bonjour"
+    #         },
+    #         "french.json": {
+    #             "hello": "bonjour"
+    #         },
+    #         "spanish.json": {
+    #             "hello": "hola",
+    #             "bye": "chau",
+    #             "what": "que",
+    #             "como": "how",
+    #             "codigo": "code"
+    #         }
+    #     }
+    #
+    #     pull_translations(write_directory=temp_dir_path)
+    #     for file_name in os.listdir(temp_dir_path):
+    #         file_path = os.path.join(temp_dir_path, file_name)
+    #         file_content = get_json(file_path)
+    #         self.assertEqual(file_content, expected_file_content[file_name])
+    #
+    #     # Cleanup
+    #     shutil.rmtree(temp_dir_path)
+    #     globals.token.value = prev_token
 
 if __name__ == '__main__':
     unittest.main()
